LayerTreeCoordinatorProxy should use uint64_t for surface key. https://bugs.webkit.org/show_bug.cgi?id=95175
GraphicsSurface tokens are of type uint64_t.
Therefore LayerTreeCoordinatorProxy must use the same type to
identify a GraphicsSurface/ShareableSurface.

Reviewed by Noam Rosenthal.

* Shared/ShareableSurface.h:
(WebKit::ShareableSurface::Handle::graphicsSurfaceToken):
* UIProcess/CoordinatedGraphics/LayerTreeCoordinatorProxy.cpp:
(WebKit::LayerTreeCoordinatorProxy::updateTileForLayer):
* UIProcess/CoordinatedGraphics/LayerTreeCoordinatorProxy.h:
(LayerTreeCoordinatorProxy):

Canonical link: https://commits.webkit.org/113054@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@126877 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit2/Shared/ShareableSurface.h
@@ -55,7 +55,7 @@ public:
         static bool decode(CoreIPC::ArgumentDecoder*, Handle&);
 
 #if USE(GRAPHICS_SURFACE)
-        uint32_t graphicsSurfaceToken() const { return m_graphicsSurfaceToken; }
+        uint64_t graphicsSurfaceToken() const { return m_graphicsSurfaceToken; }
 #endif
 
     private:
